virt-ctrl, net: Mv GetNetworkToResourceMap
Move `GetNetworkToResourceMap` and `getResourceNameForNetwork`
under `pkg/network/multus`.

Drop unit tests of `getResourceNameForNetwork` as they test
an implementation detail.
