+++ pkg/virt-controller/services/BUILD.bazel
@@ -23,6 +23,7 @@ go_library(
         "//pkg/monitoring/metrics/virt-controller:go_default_library",
         "//pkg/network/downwardapi:go_default_library",
         "//pkg/network/istio:go_default_library",
+        "//pkg/network/multus:go_default_library",
         "//pkg/network/vmispec:go_default_library",
         "//pkg/pointer:go_default_library",
         "//pkg/storage/backend-storage:go_default_library",
@@ -32,7 +33,6 @@ go_library(
         "//pkg/util/hardware:go_default_library",
         "//pkg/util/net/dns:go_default_library",
         "//pkg/virt-config:go_default_library",
-        "//pkg/virt-controller/network:go_default_library",
         "//pkg/virt-controller/watch/descheduler:go_default_library",
         "//pkg/virt-controller/watch/topology:go_default_library",
         "//pkg/virt-launcher/virtwrap/api:go_default_library",
